+++ spec/internal/db/schema.rb
@@ -15,4 +15,10 @@ ActiveRecord::Schema.define do
     t.belongs_to :user
     t.timestamps
   end
+
+  create_table(:orders, force: true) do |t|
+    t.integer :item_count
+    t.integer :total
+    t.timestamps
+  end
 end