+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Make each callback name and chain node unique as to not share nodes between models
 - Properly order callback methods in which they will be executed
 - Add sort order, instance method identifier, and symbol notation to diagram labels for better readability
+- Show validation type and attributes instead of stringified class name
 
 ## [0.1.0]
 - Initial release